remove all warnings in cargo clippy both side

+++ src/arch/aarch64/registers.rs
@@ -85,9 +85,9 @@ impl ArchReg {
             ArchReg::MsgInfo => 1,
             ArchReg::FaultIP => 34,
             ArchReg::NextIP => 32,
-            ArchReg::Msg(i) => msgRegister[*i as usize],
-            ArchReg::Frame(i) => frameRegisters[*i as usize],
-            ArchReg::GP(i) => gpRegisters[*i as usize],
+            ArchReg::Msg(i) => msgRegister[*i],
+            ArchReg::Frame(i) => frameRegisters[*i],
+            ArchReg::GP(i) => gpRegisters[*i],
             ArchReg::FaultMessage(id, index) => fault_messages[*id][*index],
         }
     }
